feat: convert to TypeScript (#6)
BREAKING CHANGE: The Lunr component has been removed. It was limited in
functionality and relied on some React anti-patterns.

If you are using a React version >= 16.8, migrating from the Lunr
component to the `useLunr` hook is highly recommended.

If you cannot convert to the hook, please lock your version of this
package to your currently installed version.